Move isValidJwtToken from provider to validator
